rename the redis cache module to `redis.py`
There's no need for redundant context, it's all bound by the scope of the "cache" package
