+++ maestro_python_client/__init__.py
@@ -3,7 +3,7 @@ Maestro client implemented in python
 """
 
 from maestro_python_client.abc.cache import AbstractCache
-from maestro_python_client.Cache.RedisCache import RedisCache
+from maestro_python_client.cache.redis import RedisCache
 from maestro_python_client.CachedClient import CachedClient
 from maestro_python_client.Client import Client, Task, TaskHistory
